# backend/src/main/java/com/moat/auth/AuthController.java
package com.moat.auth;

import com.moat.api.model.LoginRequest;
import com.moat.api.model.RegisterRequest;
import com.moat.api.model.UserResponse;
import com.moat.user.User;
import com.moat.user.UserRepository;
import jakarta.validation.Valid;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.BadCredentialsException;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.oauth2.jwt.Jwt;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;

@RestController
@RequestMapping("/auth")
public class AuthController {

    private final AuthService authService;
    private final JwtService jwtService;
    private final AuthCookieFactory cookieFactory;
    private final UserRepository userRepository;
    private final UserMapper userMapper;

    public AuthController(AuthService authService, JwtService jwtService,
                          AuthCookieFactory cookieFactory, UserRepository userRepository,
                          UserMapper userMapper) {
        this.authService = authService;
        this.jwtService = jwtService;
        this.cookieFactory = cookieFactory;
        this.userRepository = userRepository;
        this.userMapper = userMapper;
    }

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    public UserResponse register(@Valid @RequestBody RegisterRequest request) {
        return userMapper.toResponse(authService.register(request));
    }

    @PostMapping("/login")
    public ResponseEntity<UserResponse> login(@Valid @RequestBody LoginRequest request) {
        User user = authService.authenticate(request);
        String token = jwtService.issueToken(user);
        return ResponseEntity.ok()
                .header(HttpHeaders.SET_COOKIE, cookieFactory.authCookie(token).toString())
                .body(userMapper.toResponse(user));
    }

    @PostMapping("/logout")
    public ResponseEntity<Void> logout() {
        return ResponseEntity.noContent()
                .header(HttpHeaders.SET_COOKIE, cookieFactory.logoutCookie().toString())
                .build();
    }

    @GetMapping("/me")
    public UserResponse me(@AuthenticationPrincipal Jwt jwt) {
        UUID userId = UUID.fromString(jwt.getSubject());
        User user = userRepository.findById(userId)
                .orElseThrow(() -> new BadCredentialsException("User no longer exists"));
        return userMapper.toResponse(user);
    }
}